feat: implement logging error in separate file
