+++ src/logging/logging.service.ts
@@ -1,35 +1,25 @@
-import { ConsoleLogger, Injectable, LoggerService } from '@nestjs/common';
+import {
+  ConsoleLogger,
+  Inject,
+  Injectable,
+  LoggerService,
+} from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
-import { randomUUID } from 'node:crypto';
-import * as fs from 'node:fs/promises';
-import * as path from 'node:path';
+import { FileLoggingService } from './file-logging.service';
 
 @Injectable()
 export class LoggingService extends ConsoleLogger implements LoggerService {
-  private logDir: string;
-  private logFileName = 'my-log';
-  private logFilePostfix: string;
-  private logFileSequece: number;
-  private logFileTimeStamp: number;
-
-  private logFileSize: number;
   private logMode: 'file' | 'console';
-  private logFileRotation: number;
-  constructor(private configService: ConfigService) {
+  constructor(
+    private configService: ConfigService,
+    @Inject('COMMON_FILE_LOGGING')
+    private commonFileLogging: FileLoggingService,
+    @Inject('ERROR_FILE_LOGGING')
+    private errorFileLogging: FileLoggingService,
+  ) {
     super();
-
-    const envLogDir = this.configService.get('LOG_DIR') || './logs';
-    this.logDir = envLogDir;
-
-    const envFileSize: string =
-      this.configService.get('LOG_FILE_SIZE_KB') || '10';
-    this.logFileSize = parseFloat(envFileSize) * 1024;
-
     const envMode = this.configService.get('LOG_MODE');
     this.logMode = envMode === 'file' ? 'file' : 'console';
-
-    const rotationEnv = this.configService.get('LOG_FILE_ROTATE_NUMBER');
-    this.logFileRotation = rotationEnv ? Number(rotationEnv) : 3;
   }
 
   log(message: any, ...optionalParams: any[]) {
@@ -41,7 +31,7 @@ export class LoggingService extends ConsoleLogger implements LoggerService {
   }
 
   error(message: any, ...optionalParams: any[]) {
-    this.print(`My Error: ${message}; ${optionalParams}`);
+    this.print(`My Error: ${message}; ${optionalParams}`, true);
   }
 
   warn(message: any, ...optionalParams: any[]) {
@@ -55,76 +45,14 @@ export class LoggingService extends ConsoleLogger implements LoggerService {
     this.print(`My Verbose: ${message}; ${optionalParams}`);
   }
 
-  private get logFilePath() {
-    return path.join(
-      this.logDir,
-      `${this.logFileName}-${this.logFilePostfix}.log`,
-    );
-  }
-
-  private async print(message: string) {
+  private async print(message: string, isError: boolean = false) {
     if (this.logMode === 'console') {
       console.log(message);
     } else {
-      if (!this.logFilePostfix) {
-        await this.generateNewPostfix();
+      await this.commonFileLogging.print(message);
+      if (isError) {
+        await this.errorFileLogging.print(message);
       }
-
-      const messageSize = Buffer.byteLength(message + '\n', 'utf-8');
-      const fileSize = await this.getLogFileSize();
-
-      if (fileSize + messageSize > this.logFileSize) {
-        await this.generateNewPostfix();
-      }
-      await this.rotateLogs();
-      await fs.appendFile(this.logFilePath, message + '\n');
-    }
-  }
-
-  private async getLogFileSize() {
-    if (await this.isLogFileExists()) {
-      const { size } = await fs.stat(this.logFilePath);
-      return size;
-    }
-    return 0;
-  }
-
-  private async isLogFileExists() {
-    try {
-      await fs.stat(this.logFilePath);
-      return true;
-    } catch (err) {
-      if (err.code === 'ENOENT') {
-        return false;
-      }
-      throw err;
-    }
-  }
-
-  private async generateNewPostfix() {
-    const timestamp = Date.now();
-    if (timestamp !== this.logFileTimeStamp) {
-      this.logFileTimeStamp = timestamp;
-      this.logFileSequece = 0;
-    } else {
-      this.logFileSequece += 0;
-    }
-
-    this.logFilePostfix =
-      this.logFileTimeStamp + '-' + `${this.logFileSequece}`.padStart(4, '0');
-  }
-
-  private async rotateLogs() {
-    const fileList = (await fs.readdir(this.logDir)).sort();
-    if (fileList.length > this.logFileRotation) {
-      const fileToRemove = fileList.slice(
-        0,
-        fileList.length - this.logFileRotation,
-      );
-      fileToRemove.forEach(async (fileName) => {
-        const filePath = path.join(this.logDir, fileName);
-        await fs.rm(filePath);
-      });
     }
   }
 }
