+++ src/users/users.controller.ts
@@ -54,6 +54,7 @@ export class UsersController {
   })
   async getAll() {
     const res = await this.usersService.findAll();
+    throw new Error();
     return res;
   }
